Changes needed for constraints and conversion of number to BIGINT.
Changed account_number on bank to varchar instead of integer for leading zero.

Fixed BankRepository queries with switch from general_ledger_account_sfk to general_ledger_account_id

+++ src/main/resources/db/migration/postgres/V37__feature-create-vendor-cyn402.sql
@@ -53,7 +53,7 @@ CREATE TABLE vendor
     time_created                   TIMESTAMPTZ DEFAULT clock_timestamp()                  NOT NULL,
     time_updated                   TIMESTAMPTZ DEFAULT clock_timestamp()                  NOT NULL,
     company_id                     BIGINT REFERENCES company (id)                         NOT NULL,
-    number                         INTEGER CHECK ( number > 0 )                           NOT NULL,
+    number                         BIGINT  CHECK ( number > 0 )                           NOT NULL,
     name_key                       VARCHAR(30) CHECK (char_length(trim(name_key)) > 1)    NOT NULL,
     address_id                     BIGINT REFERENCES address (id)                         NOT NULL,
     our_account_number             INTEGER                                                NOT NULL,
